+++ core/data/build.gradle.kts
@@ -8,4 +8,8 @@ android {
     namespace = "com.iguana.data"
 }
 dependencies {
+
+    implementation("com.squareup.retrofit2:retrofit:2.9.0")
+    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
+
 }
