+++ core/data/src/main/java/com/iguana/data/remote/api/LoginApi.kt
@@ -1,4 +1,4 @@
-package com.iguana.login
+package com.iguana.data.remote.api
 
 import retrofit2.Response
 import retrofit2.http.Body
